requirements.txt
requests~=2.22.0
praw==6.5.1
python-dotenv==0.10.3
pillow==7.0.0
progress==1.5
SQLAlchemy~=1.3.8
typing~=3.6.4